ensure the specialized serializers loaded on include by restpack
