+++ app/serializers/project_role_serializer.rb
@@ -1,7 +1,9 @@
-class ProjectRoleSerializer < AccessControlListSerializer
+class ProjectRoleSerializer
+  include ACLSerializer
+
   attributes :id, :roles
   can_include :user_group, :resource
-  
+
   def self.key
     "project_roles"
   end
